make required changes
