+++ helpers.rb
@@ -3,6 +3,7 @@ module Helpers
     sleep 1
     puts "\n#{name} created successfully!!!"
     sleep 1
+    
   end
 
   def continue?
@@ -11,6 +12,7 @@ module Helpers
 
     if answer.downcase == 'y' || answer.downcase == 'yes' || answer == ''
       clear
+      prompt_user
     else
       exit
     end
@@ -19,6 +21,7 @@ module Helpers
   def invalid_prompt
     clear
     puts 'Incorrect selection, please try again!'
+    
     sleep 1
   end
 
@@ -30,4 +33,5 @@ module Helpers
     clear
     nil
   end
+
 end
