+++ store.rb
@@ -1,0 +1,8 @@
+# class Store
+#   attr_reader :people, :books
+
+#   def initialize
+#     @people = []
+#     @books = []
+#   end
+end
